fix: global logger
